ComputedDefaultValue, refactor tests
