+++ lib/csdl2openapi.js
@@ -2162,10 +2162,7 @@ module.exports.csdl2openapi = function (
             ] = ref("count");
           }
         }
-        if (
-          property[voc.Core.Permissions] != "Read" &&
-          !property[voc.Core.Computed]
-        ) {
+        if (property[voc.Core.Permissions] != "Read") {
           if (property.$ContainsTarget) {
             if (suffix === SUFFIX.create)
               schemaProperties[name] = schema(property, SUFFIX.create);
@@ -2186,20 +2183,26 @@ module.exports.csdl2openapi = function (
       } else {
         if (
           property[voc.Core.Permissions] == "Read" ||
-          property[voc.Core.Computed]
+          property[voc.Core.Computed] ||
+          property[voc.Core.ComputedDefaultValue]
         ) {
           let index = required.indexOf(name);
           if (index != -1) required.splice(index, 1);
-        } else {
-          if (suffix === SUFFIX.create)
-            schemaProperties[name] = schema(property, SUFFIX.create);
-          if (
-            suffix === SUFFIX.update &&
-            !isKey[name] &&
-            !property[voc.Core.Immutable]
-          )
-            schemaProperties[name] = schema(property, SUFFIX.update);
         }
+        if (
+          property[voc.Core.Permissions] == "Read" ||
+          property[voc.Core.Computed]
+        )
+          continue;
+
+        if (suffix === SUFFIX.create)
+          schemaProperties[name] = schema(property, SUFFIX.create);
+        if (
+          suffix === SUFFIX.update &&
+          !isKey[name] &&
+          !property[voc.Core.Immutable]
+        )
+          schemaProperties[name] = schema(property, SUFFIX.update);
       }
     }
