+++ lib/edm.js
@@ -344,6 +344,7 @@ function vocabularies(voc, alias) {
     Core: [
       "AcceptableMediaTypes",
       "Computed",
+      "ComputedDefaultValue",
       "DefaultNamespace",
       "Description",
       "Example",
